--- DirectX12TriangleTutorial/DirectX12Triangle/main.cpp
@@ -1,12 +1,0 @@
-#include "Engine.h"
-
-//#pragma comment(lib, "d3d12.lib")
-//#pragma comment(lib, "dxgi.lib")
-//#pragma comment(lib, "d3dcompiler.lib")
-
-int WINAPI wWinMain(HINSTANCE hInstance, HINSTANCE, PWSTR, int nCmdShow) {
-    Engine engine(hInstance, 800, 800);
-    engine.Init();
-    engine.Run();
-    return 0;
-}
